backend/requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
beartype==0.14.1
better-abc==0.0.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
einops==0.8.1
fancy-einsum==0.0.3
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
jaxtyping==0.3.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.16
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
sae-lens==6.12.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sentence-transformers==5.1.2
sentencepiece==0.2.0
sentry-sdk==2.31.0
setproctitle==1.3.6
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
torchmetrics==1.8.0
tqdm==4.67.1
transformer-lens==2.16.1
transformers==4.52.4
transformers-stream-generator==0.0.5
triton==3.3.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wadler_lindig==0.1.7
wandb==0.20.1
xxhash==3.5.0
yarl==1.20.1
streamlit==1.48.0
vl-convert-python==1.8.0
kagglehub==0.3.13
beautifulsoup4==4.13.5
umap-learn==0.5.9.post2
shap==0.48.0
openai==2.7.1
--extra-index-url=https://pypi.nvidia.com
cudf-cu12==25.8.*
dask-cudf-cu12==25.8.*
cuml-cu12==25.8.*
cugraph-cu12==25.8.*
nx-cugraph-cu12==25.8.*
cuxfilter-cu12==25.8.*
cucim-cu12==25.8.*
pylibraft-cu12==25.8.*
raft-dask-cu12==25.8.*
cuvs-cu12==25.8.*